requirements.txt
streamlit
langchain
google-generativeai
langchain-google-genai
python-dotenv
pypdf2
pandas
matplotlib
langchain-community
faiss-cpu